Add the model.git package and move git and code related model classes (ChangedFile, Diff, CodeComment) from the extraction package to model.git
