+++ src/main/java/de/uhd/ifi/se/decision/management/jira/extraction/RationaleFromCodeCommentExtractor.java
@@ -3,6 +3,7 @@ package de.uhd.ifi.se.decision.management.jira.extraction;
 import de.uhd.ifi.se.decision.management.jira.model.DecisionKnowledgeElement;
 import de.uhd.ifi.se.decision.management.jira.model.DocumentationLocation;
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
+import de.uhd.ifi.se.decision.management.jira.model.git.CodeComment;
 import de.uhd.ifi.se.decision.management.jira.model.impl.DecisionKnowledgeElementImpl;
 
 import java.util.ArrayList;
