+++ src/main/java/de/uhd/ifi/se/decision/management/jira/extraction/TangledCommitDetection.java
@@ -1,5 +1,8 @@
 package de.uhd.ifi.se.decision.management.jira.extraction;
 
+import de.uhd.ifi.se.decision.management.jira.model.git.ChangedFile;
+import de.uhd.ifi.se.decision.management.jira.model.git.Diff;
+
 /**
  * Interface for the estimation whether a {@link Diff} of {@ChangedFile}s
  * contains wrong links, i.e., is tangled.
