+++ src/main/java/de/uhd/ifi/se/decision/management/jira/model/git/ChangedFile.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.extraction;
+package de.uhd.ifi.se.decision.management.jira.model.git;
 
 import java.io.File;
 import java.util.List;
