+++ src/main/java/de/uhd/ifi/se/decision/management/jira/model/git/Diff.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.extraction;
+package de.uhd.ifi.se.decision.management.jira.model.git;
 
 import java.util.List;
